requirements.txt
# --- Deep learning core ---
torch>=2.2,<3.0
torchvision>=0.17
timm>=1.0.9
ultralytics>=8.2.0

# --- Image processing & augmentation ---
albumentations>=1.4.0
opencv-python-headless>=4.9
pillow>=10.2
numpy>=1.26,<2.1

# --- Config & data ---
pyyaml>=6.0
pandas>=2.1
scikit-learn>=1.4

# --- Inference / export ---
onnx>=1.15
onnxruntime>=1.17

# --- API stack ---
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
pydantic-settings>=2.2
python-multipart>=0.0.9
slowapi>=0.1.9

# --- Async workers ---
celery[redis]>=5.3
redis>=5.0

# --- Observability ---
prometheus-client>=0.20
structlog>=24.1

# --- Experiment tracking (optional, toggled via config) ---
wandb>=0.16
mlflow>=2.12

# --- Utils ---
tqdm>=4.66
click>=8.1
requests>=2.31
